backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
sqlalchemy
google-generativeai
yt-dlp
thefuzz
python-levenshtein
